test(systems): split funded-account waits out of the Beryl helper
eip8130 only needs wait_for_balance. Compiling start_beryl_system in that binary trips dead_code, so keep the wait helpers in a shared module the Cobalt tests can include alone.
