+++ etc/systems/tests/policy_registry.rs
@@ -1,5 +1,7 @@
 //! System tests for the policy registry precompile over Base node RPC.
 
+#[path = "common/balance.rs"]
+mod balance;
 #[path = "common/beryl.rs"]
 mod beryl;
 mod common;
